+++ components/dashboard/dashboard-overview.tsx
@@ -30,11 +30,13 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { RecentActivities } from "@/components/dashboard/recent-activities";
 import { SalesChart } from "@/components/dashboard/sales-chart";
 import { fetcher } from "@/lib/api";
+import { useRouter } from "next/navigation";
 
 export function DashboardOverview() {
   // Fetch counts for each entity
   const { data: victimsData } = useSWR("/victims", fetcher);
   const { data: casesData } = useSWR("/cases", fetcher);
+  const router = useRouter()
   const { data: suspectsData } = useSWR("/suspects", fetcher);
   const { data: officersData } = useSWR("/police-officers", fetcher);
   const { data: facilitiesData } = useSWR("/health-facilities", fetcher);
@@ -70,7 +72,7 @@ export function DashboardOverview() {
   return (
     <div className="flex flex-col gap-6 p-6 bg-gray-100 rounded-xl shadow-lg">
       <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-4">
-        <Card className="border-blue-200">
+        <Card onClick={()=> router.push("/dashboard/policePosts")} className="border-blue-200">
           <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2 bg-white">
             <CardTitle className="text-sm font-medium text-blue-900">
               Police Posts
@@ -81,7 +83,7 @@ export function DashboardOverview() {
             <div className="text-2xl font-bold text-blue-900">{postsCount}</div>
           </CardContent>
         </Card>
-        <Card className="border-blue-200">
+        <Card onClick={()=> router.push("/dashboard/police-officer")} className="border-blue-200">
           <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2 bg-white">
             <CardTitle className="text-sm font-medium text-blue-900">
               Police Officers
@@ -92,7 +94,7 @@ export function DashboardOverview() {
             <div className="text-2xl font-bold text-blue-900">{officersCount}</div>
           </CardContent>
         </Card>
-        <Card className="border-blue-200">
+        <Card onClick={()=> router.push("/dashboard/cases")} className="border-blue-200">
           <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2 bg-white">
             <CardTitle className="text-sm font-medium text-blue-900">Cases</CardTitle>
             <FileText className="h-4 w-4 text-blue-400" />
@@ -101,7 +103,7 @@ export function DashboardOverview() {
             <div className="text-2xl font-bold text-blue-900">{casesCount}</div>
           </CardContent>
         </Card>
-        <Card className="border-blue-200">
+        <Card onClick={()=> router.push("/dashboard/victims")} className="border-blue-200">
           <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2 bg-white">
             <CardTitle className="text-sm font-medium text-blue-900">Victims</CardTitle>
             <Users className="h-4 w-4 text-blue-400" />
@@ -110,7 +112,7 @@ export function DashboardOverview() {
             <div className="text-2xl font-bold text-blue-900">{victimsCount}</div>
           </CardContent>
         </Card>
-        <Card className="border-blue-200">
+        <Card onClick={()=> router.push("/dashboard/suspects")} className="border-blue-200">
           <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2 bg-white">
             <CardTitle className="text-sm font-medium text-blue-900">Suspects</CardTitle>
             <UserCheck className="h-4 w-4 text-blue-400" />
@@ -119,7 +121,7 @@ export function DashboardOverview() {
             <div className="text-2xl font-bold text-blue-900">{suspectsCount}</div>
           </CardContent>
         </Card>
-        <Card className="border-blue-200">
+        <Card onClick={()=> router.push("/dashboard/charges")} className="border-blue-200">
           <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2 bg-white">
             <CardTitle className="text-sm font-medium text-blue-900">Charges</CardTitle>
             <Gavel className="h-4 w-4 text-blue-400" />
@@ -128,7 +130,7 @@ export function DashboardOverview() {
             <div className="text-2xl font-bold text-blue-900">{chargesCount}</div>
           </CardContent>
         </Card>
-        <Card className="border-blue-200">
+        <Card onClick={()=> router.push("/dashboard/facilities")} className="border-blue-200">
           <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2 bg-white">
             <CardTitle className="text-sm font-medium text-blue-900">
               Health Facilities
@@ -139,7 +141,7 @@ export function DashboardOverview() {
             <div className="text-2xl font-bold text-blue-900">{facilitiesCount}</div>
           </CardContent>
         </Card>
-        <Card className="border-blue-200">
+        <Card onClick={()=> router.push("/dashboard/suspects")} className="border-blue-200">
           <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2 bg-white">
             <CardTitle className="text-sm font-medium text-blue-900">Examinations</CardTitle>
             <Briefcase className="h-4 w-4 text-blue-400" />
